refactor(crop): rename files
